requirements.txt
numpy
requests
tqdm
opencv-python
scikit-image
zlib
dominate
pillow
dlib
typing
fastai
setuptools
torch
torchvision